pyproject.toml
[project]
name = "aml-dashboard"
version = "0.1.0"
description = "AML Dashboard based on IBM synthetic data."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-bigquery>=3.41.0",
    "google-cloud-storage>=3.10.1",
    "kaggle>=2.0.0",
]
